+++ app/src/app/components/InfoTip/content.ts
