+++ app/src/app/components/InfoTip/index.tsx
@@ -1,0 +1,2 @@
+import {InfoTip} from './InfoTip';
+export default InfoTip;
