+++ app/src/app/components/sidebar/DataPanels.tsx
@@ -1,6 +1,6 @@
 import {Box} from '@radix-ui/themes';
 import Evaluation from '@components/sidebar/Evaluation';
-import {PopulationDataPanel} from './charts/HorizontalBarChart';
+import PopulationPanel from '@components/sidebar/PopulationPanel';
 import {Tabs} from '@radix-ui/themes';
 import Layers from './Layers';
 import React from 'react';
@@ -21,7 +21,7 @@ const defaultPanels: DataPanelSpec[] = [
   {
     title: 'population',
     label: 'Population',
-    content: <PopulationDataPanel />,
+    content: <PopulationPanel />,
   },
   {
     title: 'layers',
