+++ app/src/app/components/sidebar/Evaluation.tsx
@@ -138,7 +138,6 @@ const Evaluation: React.FC = () => {
     queryClient
   );
 
-
   const {
     unassigned,
     maxValues,
