+++ app/src/app/components/sidebar/Layers.tsx
@@ -46,7 +46,7 @@ export default function Layers() {
           parentsAreBroken && mapOptions.showBrokenDistricts ? '3' : '',
           mapOptions.lockPaintedAreas === true ? '4' : '',
           mapOptions.higlightUnassigned === true ? 'higlightUnassigned' : '',
-          mapOptions.showPopulationTooltip === true ? 'showPopulationTooltip' : ''
+          mapOptions.showPopulationTooltip === true ? 'showPopulationTooltip' : '',
         ]}
       >
         <CheckboxGroup.Item
@@ -73,14 +73,24 @@ export default function Layers() {
         >
           Highlight broken precincts
         </CheckboxGroup.Item>
-        <CheckboxGroup.Item value="higlightUnassigned" onClick={() => setMapOptions({
-          higlightUnassigned: !mapOptions.higlightUnassigned
-        })}>
+        <CheckboxGroup.Item
+          value="higlightUnassigned"
+          onClick={() =>
+            setMapOptions({
+              higlightUnassigned: !mapOptions.higlightUnassigned,
+            })
+          }
+        >
           Highlight unassigned units
         </CheckboxGroup.Item>
-        <CheckboxGroup.Item value="showPopulationTooltip" onClick={() => setMapOptions({
-          showPopulationTooltip: !mapOptions.showPopulationTooltip
-        })}>
+        <CheckboxGroup.Item
+          value="showPopulationTooltip"
+          onClick={() =>
+            setMapOptions({
+              showPopulationTooltip: !mapOptions.showPopulationTooltip,
+            })
+          }
+        >
           Show population tooltip
         </CheckboxGroup.Item>
       </CheckboxGroup.Root>
