+++ app/src/app/components/sidebar/PopulationPanel/PopulationChart/types.ts
@@ -1,0 +1,7 @@
+type TooltipInput = {
+  y: number;
+  index: number;
+  pop: number;
+  idealPopulation?: number;
+  maxPop?: number;
+};
