+++ app/src/app/components/sidebar/PopulationPanel/index.tsx
@@ -1,0 +1,2 @@
+import {PopulationPanel} from './PopulationPanel';
+export default PopulationPanel;
